filestore: optionally force cleanup of rmdirs on non-empty collections
If we get a bad collection removal (rmdir) in the filestore journal but
the collection isn't actually empty, someone was bad.  Clean up the mess.
This is a workaround for #1471.

The actual bug is somewhere else: somehow we journaled an rmdir that was
doomed to fail.  In #1471's case, we actual journal a collection_add
(adding a link to the dir) and then a delete (which always fails).

Signed-off-by: Sage Weil <sage@newdream.net>